Add pug template engine, a template & res render
Installed pug through npm. Created index.pug, a simple view to work with each page for now. Set the template engine to pug in server.js and changed the res.json to res.render in each routes get all call and to return a title and list to the page to render.