Update view to use new information in responses
Each view now has a link attached to each list item such that it loads a page with a list of all documents with the current document as it's parent.
e.g.
	The country Canada when clicked would return a page with all provinces that have a country of Canada

Each view now has a back button linking to either a page of all documents with a parent document matching the parent of the current pages parent or a page of all documents in the same collection as the parent document
e.g.
	If viewing list of communities in Newfoundland:
		Back would return page with list of provinces in Canada
	Else if viewing list of all communities:
		Back would return page with list of all provinces

Two exceptions for the back link is that countries page doesn't have one and provinces only goes back to all countries page.

Also index view was simplified to not need if statement since it isn't shared any more.