Use url to specify page to load
Check the url and see if there is a page spcified to render,
still only home to render for now.

Also wrote a function to reset the URL to end with TickList,
rather than leave the page initially rendered in the URL,
you might not stay on that page but the URL wouldn't update.

With this new URL stuff i had to fix the base value in the vite config.